Fix and upgrade Markdownlint config
As #184 has shown, the lint command was ignored on Linux due to specifics of glob expansion. Now that the glob is quoted, it works correctly across all platforms and actually reports errors.

Also, Markdownlint had issues with MD002 rule, which was deprecated in the new version in favour of MD041, so I upgraded Markdownlint to pick up these changes.

Also, MD025 was reporting violations for clash between YAML `title:` and the actual heading, which is now disabled in the config.

This supersedes #184.