Remove white flash on every page load
Previously, the root element was hidden until our script figured out whether we are in light or in dark mode. This was done to avoid seeing the light theme for a split second before the dark mode kicked in. However, it resulted in a noticable white flash on each navigation, regardless of which theme is used.

By not hiding the root element in the first place, we avoid the flash in the default case where the light theme is used. This way, there’s only a flash when the dark theme is used. It’s a trade-off, but this seems like a better experience to me, especially since this way the content is visible more quickly in all cases.