Restore proper JavaScript MIME type
We want to serve JavaScript files as `text/javascript` per the recommendation in the HTML Standard. At the same time, we want to avoid serving a non-JavaScript file whose URL happens to end with `.js` as if it were JavaScript.

This patch tightens the rules to ensure only known JavaScript files are matched.

Ref. #464, #465.