fix local cache taking 1y to expire
The serviceworker config in the generated sw.js contains the revision
of cached pages. Client-side these revisions are used by Workbox to
invalidate the local cache. The local cache strategy is
`precacheAndRoute`. As long as `sw.js` is not refreshed the local
cache is not invalidated and stale pages are served indefinetly.