Small cleanups (#740)
The description said that V8 can be executed standalone, which is incorrect. V8 always needs an embedder, and d8 is just one of the embedders of V8.

Moreover, the versions of operating systems got removed. Chrome does not support Windows 7 anymore, so V8 probably also does not support Windows 7 anymore. Since we don't keep the version numbers on v8.dev up to date, it's better to just remove them.

Additionally the first-party support for Mips was removed.